requirements_dev.txt
pip
cmake==3.21.1
wheel==0.32.1
watchdog==0.9.0
flake8==3.5.0
coverage==4.5.1
Sphinx==1.8.1
twine==1.12.1

pytest==3.8.2
pytest-runner==4.2
pytest-cov==2.7.1

matchms~=0.8.2
requests~=2.24.0
pandas~=1.1.3
pillow~=8.3.1
confuse~=1.4.0
rdkit-pypi~=2021.3.4
setuptools~=49.6.0
cryptography==3.4.7
